unittests/2022/test_days/test_day14.py
import numpy as np

from src.aoc2022 import utils
from src.aoc2022.days import day14


class TestDay14:
    def test_parse_data_to_array(self):
        # Prepare
        test_data_path = "./unittests/2022/test_data/day14_sample.txt"
        input_data = utils.get_raw_data(test_data_path)
        start = (500, 0)

        expected_map = np.array(
            [
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0],
            ]
        )

        expected_start = (6, 0)

        # Run
        actual_map, actual_start = day14.parse_data_to_array(input_data, start)

        # Assert
        assert actual_start == expected_start
        assert (actual_map == expected_map).all()

    def test_simulate_grain_of_sand(self):
        # Prepare
        test_data_path = "./unittests/2022/test_data/day14_sample.txt"
        input_data = utils.get_raw_data(test_data_path)
        starting_point = (500, 0)
        map_, start = day14.parse_data_to_array(input_data, starting_point)
        expected_sand_position = (6, 8)

        # Run
        actual_sand_position = day14.simulate_grain_of_sand(map_, start)

        # Assert
        assert actual_sand_position == expected_sand_position

    def test_add_floor_to_map(self):
        # Prepare
        test_data_path = "./unittests/2022/test_data/day14_sample.txt"
        input_data = utils.get_raw_data(test_data_path)
        starting_point = (500, 0)
        map_, start = day14.parse_data_to_array(input_data, starting_point)
        expected_start = (18, 0)

        expected_map = np.array(
            [
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0],
            ]
        )

        # Run
        actual_map, actual_start = day14.add_floor_to_map(map_, start)

        # Assert
        assert actual_start == expected_start
        assert (actual_map == expected_map).all()

    def test_simulate_sand_falling(self):
        # Prepare
        test_data_path = "./unittests/2022/test_data/day14_sample.txt"
        input_data = utils.get_raw_data(test_data_path)
        starting_point = (500, 0)
        map_, start = day14.parse_data_to_array(input_data, starting_point)

        expected_sand_count = 24
        SAND_VALUE = 2

        # Run
        actual_map = day14.simulate_sand_falling(map_, start)
        actual_sand_count = actual_map[actual_map == SAND_VALUE].shape[0]

        # Assert
        assert actual_sand_count == expected_sand_count

    def test_simulate_sand_falling_with_floor(self):
        # Prepare
        test_data_path = "./unittests/2022/test_data/day14_sample.txt"
        input_data = utils.get_raw_data(test_data_path)
        starting_point = (500, 0)
        map_, start = day14.parse_data_to_array(input_data, starting_point)
        floor_map, floor_start = day14.add_floor_to_map(map_, start)

        expected_sand_count = 93
        SAND_VALUE = 2

        # Run
        actual_map = day14.simulate_sand_falling(floor_map, floor_start)
        actual_sand_count = actual_map[actual_map == SAND_VALUE].shape[0]

        # Assert
        assert actual_sand_count == expected_sand_count

    def test_render_map(self):
        # Prepare
        test_data_path = "./unittests/2022/test_data/day14_sample.txt"
        input_data = utils.get_raw_data(test_data_path)
        starting_point = (500, 0)
        map_, start = day14.parse_data_to_array(input_data, starting_point)
        sand_map = day14.simulate_sand_falling(map_, start)

        expected_render = """......S...
..........
......o...
.....ooo..
....#ooo##
...o#ooo#.
..###ooo#.
....oooo#.
.o.ooooo#.
#########."""

        # Run
        actual_render = day14.render_map(sand_map, start)

        # Assert
        assert actual_render == expected_render
